+++ test/where_test.rb
@@ -78,7 +78,7 @@ class WhereTest < Minitest::Test
 
   def test_regexp
     store_names ["Product A"]
-    assert_search "*", ["Product A"], where: {name: /Pro.+/}
+    assert_search "*", ["Product A"], where: {name: /\APro.+\z/}
   end
 
   def test_alternate_regexp
@@ -88,7 +88,22 @@ class WhereTest < Minitest::Test
 
   def test_special_regexp
     store_names ["Product <A>", "Item <B>"]
-    assert_search "*", ["Product <A>"], where: {name: /Pro.+<.+/}
+    assert_search "*", ["Product <A>"], where: {name: /\APro.+<.+\z/}
+  end
+
+  # regular expressions are always anchored in ES
+  def test_regexp_not_anchored
+    store_names ["abcde"]
+    assert_search "*", [], where: {name: /abcd/}
+    assert_search "*", ["abcde"], where: {name: /abcde/}
+    assert_search "*", ["abcde"], where: {name: /\Aabcde\z/}
+  end
+
+  def test_regexp_case
+    store_names ["abcde"]
+    assert_search "*", [], where: {name: /\AABCDE\z/}
+    # flags don't work
+    assert_search "*", [], where: {name: /\AABCDE\z/i}
   end
 
   def test_prefix