+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 4.0.3 [unreleased]
+
+- Added warnings for certain regular expressions
+- Fixed anchored regular expressions
+
 ## 4.0.2
 
 - Added block form of `scroll`